+++ extensions/hidden-icons/CHANGELOG.md
@@ -1,5 +1,5 @@
 # Hidden Icons Changelog
 
-## [Added Hidden Icons] - 2022-04-30
+## [Added Hidden Icons] - 2022-04-26
 - Release of Hidden Icons version 1.0
-- **Find out Hidden Icon**: Find out the hidden icon in 30 seconds.
+- **Find Hidden Icon**: A casual game where you need to find a hidden icon..